no spring security, everything else works, frontend almost dome

+++ src/main/java/pl/coderslab/spring_boot_project/model/TaskDto.java
@@ -11,7 +11,7 @@ public class TaskDto {
 
     Long taskId;
     String name;
-    int daysToDeadline;
+    Long daysToDeadline;
     String status;
     String lastImageSource;
 }
